organizing and adding tests
